podman_remote/requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
paramiko==2.9.1
pycparser==2.21
PyNaCl==1.4.0
PyYAML==6.0
six==1.16.0
